+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 ### Internal
 
 - [Docs] Update download count image for Minkowski² (#4474 by: Clayell; reviewed: HebaruSan)
+- [Netkan] Fix x_netkan_epoch not applied before unreliable server check (#4477 by: HebaruSan)
 
 ## v1.36.2 (Politas)